+++ lib/src/classes/multi_point.dart
@@ -1,6 +1,4 @@
-import 'dart:convert';
-
-import '../../geojson_vi.dart';
+part of geojson_vi;
 
 /// A class representing a GeoJSON MultiPoint.
 ///